Renamed StateInfo.js to Router.js
